Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    int A,B;
+    float A,B;
     scanf("%f","%f",A,B);
     printf("Product: %f", A*B);
     return 0;
